+++ src/test/sh/org/apache/aurora/e2e/check-fs.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+echo "Checking volume mounts..."
+if [ ! -f "/etc/rsyslog.d.container/50-default.conf" ]; then
+    echo "Mounted file was not found";
+    exit 1
+fi