+++ src/test/sh/org/apache/aurora/e2e/run-server.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 echo "Starting up server..."
 while true
 do